requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.4
gunicorn==22.0.0
alembic==1.13.3
